feat: remove legacy gemini agent and provider adapter
