--- apps/server/src/api/services/chat-service.ts
@@ -1,297 +1,0 @@
-/**
- * @license
- * Copyright 2025 BrowserOS
- * SPDX-License-Identifier: AGPL-3.0-or-later
- */
-
-import { mkdir } from 'node:fs/promises'
-import path from 'node:path'
-import { LLM_PROVIDERS } from '@browseros/shared/schemas/llm'
-import { MCPServerConfig } from '@google/gemini-cli-core'
-import type { HonoSSEStream } from '../../agent/provider-adapter/types'
-import type { SessionManager } from '../../agent/session'
-import type { ProviderConfig, ResolvedAgentConfig } from '../../agent/types'
-import type { Browser } from '../../browser/browser'
-import { INLINED_ENV } from '../../env'
-import {
-  fetchBrowserOSConfig,
-  getLLMConfigFromProvider,
-} from '../../lib/clients/gateway'
-import type { KlavisClient } from '../../lib/clients/klavis/klavis-client'
-import { logger } from '../../lib/logger'
-import {
-  detectMcpTransport,
-  type McpTransportType,
-} from '../../lib/mcp-transport-detect'
-import type { BrowserContext, ChatRequest } from '../types'
-
-interface McpServerOptions {
-  url: string
-  transport: McpTransportType
-  headers?: Record<string, string>
-  trust?: boolean
-}
-
-function createMcpServerConfig(options: McpServerOptions): MCPServerConfig {
-  return new MCPServerConfig(
-    undefined,
-    undefined,
-    undefined,
-    undefined,
-    options.transport === 'sse' ? options.url : undefined,
-    options.transport === 'streamable-http' ? options.url : undefined,
-    options.headers,
-    undefined,
-    undefined,
-    options.trust,
-  )
-}
-
-export interface ChatServiceDeps {
-  sessionManager: SessionManager
-  klavisClient: KlavisClient
-  executionDir: string
-  mcpServerUrl: string
-  browserosId?: string
-  browser: Browser
-}
-
-export class ChatService {
-  constructor(private deps: ChatServiceDeps) {}
-
-  async processMessage(
-    request: ChatRequest,
-    rawStream: HonoSSEStream,
-    abortSignal: AbortSignal,
-  ): Promise<void> {
-    const { sessionManager } = this.deps
-
-    const providerConfig = await this.resolveProviderConfig(request)
-    logger.debug('Provider config resolved', {
-      provider: providerConfig.provider,
-      model: providerConfig.model,
-      hasUpstreamProvider: !!providerConfig.upstreamProvider,
-    })
-
-    const mcpServers = await this.buildMcpServers(
-      request.conversationId,
-      request.browserContext,
-    )
-    logger.debug('MCP servers built', {
-      serverCount: Object.keys(mcpServers).length,
-      servers: Object.keys(mcpServers),
-    })
-
-    const sessionExecutionDir = await this.resolveSessionDir(request)
-
-    const agentConfig: ResolvedAgentConfig = {
-      conversationId: request.conversationId,
-      provider: providerConfig.provider,
-      model: providerConfig.model,
-      apiKey: providerConfig.apiKey,
-      baseUrl: providerConfig.baseUrl,
-      upstreamProvider: providerConfig.upstreamProvider,
-      resourceName: providerConfig.resourceName,
-      region: providerConfig.region,
-      accessKeyId: providerConfig.accessKeyId,
-      secretAccessKey: providerConfig.secretAccessKey,
-      sessionToken: providerConfig.sessionToken,
-      contextWindowSize: request.contextWindowSize,
-      userSystemPrompt: request.userSystemPrompt,
-      sessionExecutionDir,
-      supportsImages: request.supportsImages,
-      chatMode: request.mode === 'chat',
-      isScheduledTask: request.isScheduledTask,
-    }
-
-    const browserContext = await this.resolvePageIds(request.browserContext)
-
-    const session = await sessionManager.getOrCreate(agentConfig, mcpServers)
-    await session.agent.execute(
-      request.message,
-      rawStream,
-      abortSignal,
-      browserContext,
-      request.previousConversation,
-    )
-  }
-
-  private async resolveProviderConfig(
-    request: ChatRequest,
-  ): Promise<ProviderConfig> {
-    if (request.provider === LLM_PROVIDERS.BROWSEROS) {
-      const configUrl = INLINED_ENV.BROWSEROS_CONFIG_URL
-      if (!configUrl) {
-        throw new Error(
-          'BROWSEROS_CONFIG_URL environment variable is required for BrowserOS provider',
-        )
-      }
-
-      logger.info('Fetching BrowserOS config', {
-        configUrl,
-        browserosId: this.deps.browserosId,
-      })
-
-      const browserosConfig = await fetchBrowserOSConfig(
-        configUrl,
-        this.deps.browserosId,
-      )
-      const llmConfig = getLLMConfigFromProvider(browserosConfig, 'default')
-
-      logger.info('Using BrowserOS config', {
-        model: llmConfig.modelName,
-        baseUrl: llmConfig.baseUrl,
-        upstreamProvider: llmConfig.providerType,
-      })
-
-      return {
-        provider: request.provider,
-        model: llmConfig.modelName,
-        apiKey: llmConfig.apiKey,
-        baseUrl: llmConfig.baseUrl,
-        upstreamProvider: llmConfig.providerType,
-      }
-    }
-
-    return {
-      provider: request.provider,
-      model: request.model,
-      apiKey: request.apiKey,
-      baseUrl: request.baseUrl,
-      resourceName: request.resourceName,
-      region: request.region,
-      accessKeyId: request.accessKeyId,
-      secretAccessKey: request.secretAccessKey,
-      sessionToken: request.sessionToken,
-    }
-  }
-
-  private async buildMcpServers(
-    conversationId: string,
-    browserContext?: BrowserContext,
-  ): Promise<Record<string, MCPServerConfig>> {
-    const { klavisClient, mcpServerUrl, browserosId } = this.deps
-    const servers: Record<string, MCPServerConfig> = {}
-
-    if (mcpServerUrl) {
-      servers['browseros-mcp'] = createMcpServerConfig({
-        url: mcpServerUrl,
-        transport: 'streamable-http',
-        headers: {
-          Accept: 'application/json, text/event-stream',
-          'X-BrowserOS-Scope-Id': conversationId,
-        },
-        trust: true,
-      })
-    }
-
-    if (browserosId && browserContext?.enabledMcpServers?.length) {
-      try {
-        const result = await klavisClient.createStrata(
-          browserosId,
-          browserContext.enabledMcpServers,
-        )
-        servers['klavis-strata'] = createMcpServerConfig({
-          url: result.strataServerUrl,
-          transport: 'streamable-http',
-          trust: true,
-        })
-        logger.info('Added Klavis Strata MCP server', {
-          browserosId: browserosId.slice(0, 12),
-          servers: browserContext.enabledMcpServers,
-        })
-      } catch (error) {
-        logger.error('Failed to create Klavis Strata MCP server', {
-          browserosId: browserosId?.slice(0, 12),
-          servers: browserContext.enabledMcpServers,
-          error: error instanceof Error ? error.message : String(error),
-        })
-      }
-    }
-
-    if (browserContext?.customMcpServers?.length) {
-      const customServers = browserContext.customMcpServers
-      const transports = await Promise.all(
-        customServers.map((server) => detectMcpTransport(server.url)),
-      )
-      for (let i = 0; i < customServers.length; i++) {
-        const server = customServers[i]
-        const transport = transports[i]
-        servers[`custom-${server.name}`] = createMcpServerConfig({
-          url: server.url,
-          transport,
-          trust: true,
-        })
-        logger.info('Added custom MCP server', {
-          name: server.name,
-          url: server.url,
-          transport,
-        })
-      }
-    }
-
-    return servers
-  }
-
-  private async resolveSessionDir(request: ChatRequest): Promise<string> {
-    let dir: string
-    let userProvided: boolean
-
-    if (request.userWorkingDir) {
-      dir = request.userWorkingDir
-      userProvided = true
-    } else {
-      // create new session dir for this conversationId
-      dir = path.join(
-        this.deps.executionDir,
-        'sessions',
-        request.conversationId,
-      )
-      userProvided = false
-    }
-
-    await mkdir(dir, { recursive: true })
-    logger.info('Session directory resolved', { dir, userProvided })
-
-    return dir
-  }
-
-  // Browser context arrives with Chrome tab IDs, but tools expect internal page IDs.
-  // Resolve the mapping upfront so the agent's first navigation doesn't fail.
-  private async resolvePageIds(
-    browserContext?: BrowserContext,
-  ): Promise<BrowserContext | undefined> {
-    if (!browserContext) return undefined
-
-    const tabIds: number[] = []
-    if (browserContext.activeTab) tabIds.push(browserContext.activeTab.id)
-    if (browserContext.selectedTabs) {
-      for (const tab of browserContext.selectedTabs) tabIds.push(tab.id)
-    }
-    if (browserContext.tabs) {
-      for (const tab of browserContext.tabs) tabIds.push(tab.id)
-    }
-
-    if (tabIds.length === 0) return browserContext
-
-    const tabToPage = await this.deps.browser.resolveTabIds(tabIds)
-
-    const addPageId = (tab: { id: number; url?: string; title?: string }) => ({
-      ...tab,
-      pageId: tabToPage.get(tab.id),
-    })
-
-    logger.debug('Resolved tab IDs to page IDs', {
-      mapping: Object.fromEntries(tabToPage),
-    })
-
-    return {
-      ...browserContext,
-      activeTab: browserContext.activeTab
-        ? addPageId(browserContext.activeTab)
-        : undefined,
-      selectedTabs: browserContext.selectedTabs?.map(addPageId),
-      tabs: browserContext.tabs?.map(addPageId),
-    }
-  }
-}
