+++ apps/server/src/api/services/sdk/chat.ts
@@ -8,8 +8,8 @@
 
 import type { BrowserContext } from '@browseros/shared/schemas/browser-context'
 import type { LLMConfig } from '@browseros/shared/schemas/llm'
+import type { UIMessageStreamEvent } from '@browseros/shared/schemas/ui-stream'
 import { createParser, type EventSourceMessage } from 'eventsource-parser'
-import type { UIMessageStreamEvent } from '../../../agent/provider-adapter/ui-message-stream'
 import { SdkError } from './types'
 
 export interface ExecuteActionOptions {
