+++ apps/server/src/lib/clients/llm/test-provider.ts
@@ -1,0 +1,67 @@
+/**
+ * @license
+ * Copyright 2025 BrowserOS
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+import { TIMEOUTS } from '@browseros/shared/constants/timeouts'
+import type { LLMConfig } from '@browseros/shared/schemas/llm'
+import { generateText } from 'ai'
+import { resolveLLMConfig } from './config'
+import { createLLMProvider } from './provider'
+
+export interface ProviderTestConfig extends LLMConfig {
+  model: string
+  upstreamProvider?: string
+}
+
+export interface ProviderTestResult {
+  success: boolean
+  message: string
+  responseTime?: number
+}
+
+const TEST_PROMPT = "Respond with exactly: 'ok'"
+
+export async function testProviderConnection(
+  config: ProviderTestConfig,
+): Promise<ProviderTestResult> {
+  const startTime = performance.now()
+
+  try {
+    const resolvedConfig = await resolveLLMConfig(config)
+    const model = createLLMProvider(resolvedConfig)
+    const response = await generateText({
+      model,
+      messages: [{ role: 'user', content: TEST_PROMPT }],
+      abortSignal: AbortSignal.timeout(TIMEOUTS.TEST_PROVIDER),
+    })
+
+    const responseTime = Math.round(performance.now() - startTime)
+    const text = response.text
+
+    if (text) {
+      const preview = text.length > 100 ? `${text.slice(0, 100)}...` : text
+      return {
+        success: true,
+        message: `Connection successful. Response: "${preview}"`,
+        responseTime,
+      }
+    }
+
+    return {
+      success: true,
+      message: 'Connection successful. Provider responded.',
+      responseTime,
+    }
+  } catch (error) {
+    const responseTime = Math.round(performance.now() - startTime)
+    const errorMessage = error instanceof Error ? error.message : String(error)
+
+    return {
+      success: false,
+      message: `[${config.provider}] ${errorMessage}`,
+      responseTime,
+    }
+  }
+}
